# API/IMExApi/IMExApi/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "IMExApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Equipment": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Equipment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DTOEquipment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Equipment/{Id}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Equipment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DTOEquipment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DTOUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/{Id}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DTOUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DTOEquipment": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "id_OPU": {
          "format": "uuid",
          "type": "string"
        },
        "opu": {
          "type": "string"
        },
        "id_Platform": {
          "format": "uuid",
          "type": "string"
        },
        "platform": {
          "type": "string"
        },
        "id_LocationOne": {
          "format": "uuid",
          "type": "string"
        },
        "locationOne": {
          "type": "string"
        },
        "id_LocationTwo": {
          "format": "uuid",
          "type": "string"
        },
        "locationTwo": {
          "type": "string"
        },
        "id_HACDwgNo": {
          "format": "uuid",
          "type": "string"
        },
        "hacDwgNo": {
          "type": "string"
        },
        "id_Coordinate": {
          "format": "uuid",
          "type": "string"
        },
        "coordinate": {
          "type": "string"
        },
        "id_Zone": {
          "format": "uuid",
          "type": "string"
        },
        "zone": {
          "type": "string"
        },
        "id_Purpose": {
          "format": "uuid",
          "type": "string"
        },
        "purpose": {
          "type": "string"
        },
        "id_Manufacturer": {
          "format": "uuid",
          "type": "string"
        },
        "manufacturer": {
          "type": "string"
        },
        "id_Model": {
          "format": "uuid",
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "id_SerialNo": {
          "format": "uuid",
          "type": "string"
        },
        "serialNo": {
          "type": "string"
        },
        "id_IECExCertificateAuthority": {
          "format": "uuid",
          "type": "string"
        },
        "iecExCertificateAuthority": {
          "type": "string"
        },
        "id_IECExCertificateCategory": {
          "format": "uuid",
          "type": "string"
        },
        "iecExCertificateCategory": {
          "type": "string"
        },
        "id_IECExCertificateNo": {
          "format": "uuid",
          "type": "string"
        },
        "iecExCertificateNo": {
          "type": "string"
        },
        "id_IECExCertificateDetail": {
          "format": "uuid",
          "type": "string"
        },
        "iecExCertificateDetail": {
          "type": "string"
        },
        "id_ElectricalRating": {
          "format": "uuid",
          "type": "string"
        },
        "electricalRating": {
          "type": "string"
        },
        "id_Discipline": {
          "format": "uuid",
          "type": "string"
        },
        "discipline": {
          "type": "string"
        },
        "id_ProtectionType": {
          "format": "uuid",
          "type": "string"
        },
        "protectionType": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "tagNo": {
          "type": "string"
        },
        "active": {
          "type": "boolean"
        },
        "id_CreatedBy": {
          "format": "uuid",
          "type": "string"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "id_UpdatedBy": {
          "format": "uuid",
          "type": "string"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "id_DeletedBy": {
          "format": "uuid",
          "type": "string"
        },
        "dateDeleted": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
    "DTOUser": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "id_OPU": {
          "format": "uuid",
          "type": "string"
        },
        "opu": {
          "type": "string"
        },
        "id_Role": {
          "format": "uuid",
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "dateValidStart": {
          "format": "date-time",
          "type": "string"
        },
        "dateValidEnd": {
          "format": "date-time",
          "type": "string"
        },
        "active": {
          "type": "boolean"
        },
        "id_CreatedBy": {
          "format": "uuid",
          "type": "string"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "id_UpdatedBy": {
          "format": "uuid",
          "type": "string"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "id_DeletedBy": {
          "format": "uuid",
          "type": "string"
        },
        "dateDeleted": {
          "format": "date-time",
          "type": "string"
        }
      }
    }
  }
}